# docs/swagger.yaml
basePath: /
definitions:
  notes.Note:
    properties:
      created_at:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      nid:
        type: string
      status:
        type: integer
      title:
        type: string
      uid:
        type: string
    type: object
  users.User:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      uid:
        type: string
    type: object
  users.UserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@todolist.com
    name: API Support
  description: API для управления задачами и заметками
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ToDoList API
  version: "1.0"
paths:
  /notes/add:
    post:
      consumes:
      - application/json
      description: Create a new note for authenticated user
      parameters:
      - description: Note data
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/notes.Note'
      produces:
      - application/json
      responses:
        "201":
          description: Note successfully created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid JSON body
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Note creation conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create new note
      tags:
      - notes
  /notes/del/{id}:
    delete:
      description: Delete note by ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Note successfully deleted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid note ID or note not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete note
      tags:
      - notes
  /notes/list:
    get:
      description: Get list of all notes for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of notes
          schema:
            items:
              $ref: '#/definitions/notes.Note'
            type: array
        "204":
          description: No notes found
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all notes
      tags:
      - notes
  /notes/list/{id}:
    get:
      description: Get note details by note ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Note details
          schema:
            $ref: '#/definitions/notes.Note'
        "204":
          description: Note not found
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid note ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get note by ID
      tags:
      - notes
  /notes/upd/{id}:
    put:
      consumes:
      - application/json
      description: Update existing note by ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      - description: Note data to update
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/notes.Note'
      produces:
      - application/json
      responses:
        "200":
          description: Note successfully updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid JSON body or note ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Note not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update note
      tags:
      - notes
  /users/del/{id}:
    delete:
      description: Delete user account by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User successfully deleted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/users.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - users
  /users/profile:
    get:
      description: Get list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/users.User'
            type: array
        "202":
          description: No users found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all users
      tags:
      - users
  /users/profile/{id}:
    get:
      description: Get user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.User'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully registered
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: User already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register new user
      tags:
      - users
  /users/upd/{id}:
    put:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.User'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update user
      tags:
      - users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'JWT token for authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"